+++ src/main/java/gift/kakao/login/controller/KakaoLoginController.java
@@ -1,30 +1,24 @@
 package gift.kakao.login.controller;
 
-import gift.kakao.login.service.KakaoLoginService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RequestMapping;
 
-@RestController
-@Tag(name = "kakao 로그인")
-public class KakaoLoginController {
+@Controller
+@RequestMapping("/login")
+public class KakaoLoginViewController {
+    @Value("${kakao.client-id}")
+    private String clientId;
 
-    private final KakaoLoginService kakaoLoginService;
+    @Value("${kakao.redirect-uri}")
+    private String redirectUri;
 
-    public KakaoLoginController(KakaoLoginService kakaoLoginService) {
-        this.kakaoLoginService = kakaoLoginService;
-    }
-
-    @GetMapping("")
-    @Operation(summary = "kakao 로그인, Oauth callback 처리")
-    public ResponseEntity<String> handleOAuthCallback(
-        @RequestParam(value = "code") String code) {
-        String jwtAccessToken = kakaoLoginService.getAccessToken(code);
-        String kakaoEmail = kakaoLoginService.getUserInfo(jwtAccessToken);
-        return new ResponseEntity<>(jwtAccessToken, HttpStatus.OK);
+    @GetMapping
+    public String showLoginPageTest(Model model){
+        model.addAttribute("restApiKey", clientId);
+        model.addAttribute("redirectUri", redirectUri);
+        return "kakao_login";
     }
 }
