--- src/main/java/gift/kakao/login/controller/KakaoLoginViewController.java
@@ -1,24 +1,0 @@
-package gift.kakao.login.controller;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/login")
-public class KakaoLoginViewController {
-    @Value("${kakao.client-id}")
-    private String clientId;
-
-    @Value("${kakao.redirect-uri}")
-    private String redirectUri;
-
-    @GetMapping
-    public String showLoginPageTest(Model model){
-        model.addAttribute("restApiKey", clientId);
-        model.addAttribute("redirectUri", redirectUri);
-        return "kakao_login";
-    }
-}
