+++ src/main/resources/templates/kakao_login.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Title</title>
+</head>
+<body>
+<button id="login-kakao-btn" th:attr="data-rest-api-key=${restApiKey}, data-redirect-uri=${redirectUri}" onclick="redirectToKakao()">
+  카카오로 로그인하기
+</button>
+
+<script th:inline="javascript">
+  function redirectToKakao() {
+    var restApiKey = /*[[${restApiKey}]]*/ '';
+    var redirectUri = /*[[${redirectUri}]]*/ '';
+    var url = 'https://kauth.kakao.com/oauth/authorize?scope=talk_message&response_type=code&client_id=' + restApiKey + '&redirect_uri=' + redirectUri;
+    location.href = url;
+  }
+</script>
+</body>
+</html>
